Update query.py
Modifies the spectral_deconvolution function. The reference spectrum is first grouped, and then matches in the raw extracted AIF are found and assigned tags depending on their alignment status to the precursor ion. The remaining AIF peaks are then grouped and assessed for their alignment to the precursor ion to construct the complete "pseudo-MS/MS" spectrum. Method adapted from MetaboAnnotatoR (Ebbels et al., 2022).