Update query.py
Deleted unused function (reference_match) and modified composite scoring metric so that it is now the sum of the mz similarity, cosine similarity, and CCS similarity scores (theoretical maximum of 300; each is scaled to 100). The output Excel file should now display both the raw composite score for each potential match and its percentage out of the theoretical maximum of 300.